pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amalgam"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "pyarrow>=16.0",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "pytest>=8.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*"]
